RepositoryManager and simple authentication
A work-in-progress RepositoryManager was added. So far it's completely
untested. It additionally needs to be able to authenticate when working
with a remote git repository.

Authentication so far only checks if an Authorization header is present.
This obviously needs to be enhanced once we can read entity-keys from
the repository.