HackerController is at a useable state. Work on EntityController started
HackerController is working and under test. Tests for directory
traversal fail. The implementation needs to be hardened to withstand
at least the basic attacks from the test.

Implementation of EntityController was started and correct calls could
work. No tests were written so far, so the implementation could be total
trash as well.